Attempt to improve fluid tank and nether brick furnace in survival mode + Patchouli Improvement

+++ src/main/java/com/khanhpham/tothemoon/utils/multiblock/MultiblockEntity.java
@@ -1,6 +1,7 @@
 package com.khanhpham.tothemoon.utils.multiblock;
 
 import com.khanhpham.tothemoon.core.abstracts.ItemCapableBlockEntity;
+import com.khanhpham.tothemoon.utils.helpers.ModUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
@@ -14,17 +15,25 @@ public abstract class MultiblockEntity extends ItemCapableBlockEntity {
         super(pType, pWorldPosition, pBlockState, inventorySize);
     }
 
-    public void setMultiblock(@Nullable Multiblock multiblock) {
-        //if already null, no need to do anything
-        if (multiblock == null) {
-            this.onRemove();
-        }
-        this.multiblock = multiblock;
+    public void removeMultiblock() {
+        this.setMultiblock(null);
     }
 
     @Nullable
     public Multiblock getMultiblock() {
         return multiblock;
+    }
+
+    public void setMultiblock(@Nullable Multiblock multiblock) {
+        if (multiblock == null) {
+            if (this.multiblock != null) {
+                this.onRemove();
+                this.multiblock = null;
+            }
+            return;
+        }
+
+        this.multiblock = multiblock;
     }
 
     public void onRemove() {
